Start SystemD unit for client connections on Xenial and Jessie
On Ubuntu Xenial and Debian Jessie, starting the 'openvpn' service
doesn't connect a client the 'openvpn@<connection name>' service
needs to be started, where <connection name> is the name of the config
file in /etc/openvpn, minus the .conf extension.

I have not tested that this works on Jessie, but the package seems
to share the same form as the one on Xenial. This fix may be applicable
for other platforms.

Fixes #35